v-54

+++ app/Controllers/Http/BookrecordsController.ts
@@ -6,7 +6,7 @@ import Env from '@ioc:Adonis/Core/Env'
 import BadRequestException from 'App/Exceptions/BadRequestException'
 import validations from 'App/Services/Validations/validations'
 import BadRequest from 'App/Exceptions/BadRequestException'
-
+import Typebook from 'App/Models/Typebook'
 const fileRename = require('../../Services/fileRename/fileRename')
 
 export default class BookrecordsController {
@@ -200,7 +200,7 @@ export default class BookrecordsController {
   public async destroyManyBookRecords({ auth, request, response }: HttpContextContract) {
 
     const { companies_id } = await auth.use('api').authenticate()
-    const { typebooks_id, Book, startCod, endCod } = request.only(['typebooks_id', 'Book', 'startCod', 'endCod'])
+    const { typebooks_id, Book, startCod, endCod, deleteImages } = request.only(['typebooks_id', 'Book', 'startCod', 'endCod', 'deleteImages'])
 
     let query = '1 = 1'
 
@@ -212,28 +212,34 @@ export default class BookrecordsController {
         query += ` and book=${Book} `
       }
 
+
       if (startCod != undefined && endCod != undefined && startCod > 0 && endCod > 0)
         query += ` and cod>=${startCod} and cod <=${endCod} `
 
+
       try {
-
         //deletar imagem no gdrive
-        const listOfImagesToDeleteGDrive = await Indeximage
-          .query()
-          .preload('typebooks')
-          .whereIn("bookrecords_id",
-            Database.from('bookrecords')
-              .select('id')
-              .where('typebooks_id', '=', typebooks_id)
-              .andWhere('companies_id', '=', companies_id)
-              .whereRaw(query))
+        if (deleteImages) {
+          console.log("entrei no deleteImages")
+          const listOfImagesToDeleteGDrive = await Indeximage
+            .query()
+            .preload('typebooks')
+            .whereIn("bookrecords_id",
+              Database.from('bookrecords')
+                .select('id')
+                .where('typebooks_id', '=', typebooks_id)
+                .andWhere('companies_id', '=', companies_id)
+                .whereRaw(query))
 
-        if (listOfImagesToDeleteGDrive.length > 0) {
-          var file_name = listOfImagesToDeleteGDrive.map(function (item) {
-            return { file_name: item.file_name, path: item.typebooks.path }   //retorna o item original elevado ao quadrado
-          });
-          fileRename.deleteFile(file_name)
+          if (listOfImagesToDeleteGDrive.length > 0) {
+            var file_name = listOfImagesToDeleteGDrive.map(function (item) {
+              return { file_name: item.file_name, path: item.typebooks.path }   //retorna o item original elevado ao quadrado
+            });
+            fileRename.deleteFile(file_name)
+          }
+
         }
+
 
         const dataIndexImages = await Indeximage
           .query()
@@ -244,6 +250,7 @@ export default class BookrecordsController {
               .where('typebooks_id', '=', typebooks_id)
               .andWhere('companies_id', '=', companies_id)
               .whereRaw(query))
+
 
         const data = await Bookrecord
           .query()
@@ -260,12 +267,6 @@ export default class BookrecordsController {
 
     }
   }
-
-
-
-
-
-
 
   //Cria uma linha 
   public async createorupdatebookrecords({ auth, request, response }) {
@@ -501,7 +502,46 @@ export default class BookrecordsController {
   }
 
 
+  public async indeximagesinitial({ auth, params }: HttpContextContract) {
+    const authenticate = await auth.use('api').authenticate()
 
+    let listFiles
+    try {
+      const foldername = await Typebook.query().where("companies_id", "=", authenticate.companies_id).andWhere("id", "=", params.typebooks_id).first()
+      listFiles = await fileRename.indeximagesinitial(foldername, authenticate.companies_id)
+      // console.log("entrei>>>>>>", listFiles.bookRecord)
+    } catch (error) {
+      return error
+    }
+
+    for (const item of listFiles.bookRecord) {
+      try {
+        await Bookrecord.create(item)
+      } catch (error) {
+
+        console.log("ERRO BOOKRECORD::", error)
+      }
+    }
+
+    for (const item of listFiles.indexImages) {
+      try {
+        await Indeximage.create(item)
+      } catch (error) {
+        console.log("ERRO indeximage::", error)
+
+      }
+    }
+
+    // await Bookrecord.createMany(listFiles.bookRecord)
+    // console.log("executei bookrecord")
+    // await Indeximage.createMany(listFiles.indexImages)
+    // console.log("executei indeximages")
+    return "sucesso!!"
+
+
+
+
+  }
 
 
   //********************************************************* */
